requirements.txt
openai>=1.0.0
pypdf>=3.17.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1 
